Proposed implementation
A lower level implementation for our SDK. We can update session to live on top of this new MQTT implementation.
https://docs.google.com/document/d/1Zy6FGA-p7sur8aH-WZYx-DN_5Y2GgME4DjtFdAjYmWU/edit?usp=sharing
More iteration

Basic MQTT client with minimal implementation for use with for use with Mode's MQTTD. There's nothing Mode specific, but it's only been implemented for the Mode feature set.