Proposed implementation
https://docs.google.com/document/d/1Zy6FGA-p7sur8aH-WZYx-DN_5Y2GgME4DjtFdAjYmWU/edit?usp=sharing

This is the reimplementation of the device-sdk. There are still a few pieces left, but the examples are running, so it's time to get some eyes on this.

o MQTT client rewritten. It requires more work for the user, and more responsibility has been pushed to the user as well.
o Rest code is basically untouched. 
o echo examples re-written, ddp examples removed. 
o removed session, kv_cache
o packaged renamed from mode to mode_client.

Items remaining:
o Update the rest code to match the style of the mqtt stuff
o Remove the auth stuff for the tests.
o Clean up the examples. It maybe useful to use some of the code as helpers, since we removed a lot of the nice stuff from session, but we don't want to reimplement it all the time.