v3 of the module
To avoid confusion, since this is a rewrite, we will keep development on the v2 module simultaneously. This allows us to have a master branch were we can easily see both v2 and v3 modules.
Tweaks to run with v3

Most of the code was v3, so just a few more changes to get all the versions correct.