API update, fix, and race cleanup
useTLS
useTLS wasn't working because it always returned false. There was a package level variable in v2, but, because it was a named return value, it used that, so it was always false. Fixed that, and wrote a test to check all the overrides by options.
 
Added Unsubscribe
Added Unsubscribe to the API. This is needed for SDS.  Although this is blocking, you can call it from two goroutines, so we use another channel so you can do both at the same time.

Locking
race fix: while running 'with test' --race, some issues came up. Some in test code, and one minor status, so, fixed this so it reduces the noise to find serious race conditions.