check to see if the target is nil
We want to check at the connection level to see if we are connected. However, the connection itself may be nil, so added that check.
added test for nil mqttConn check

since all the code runs through two functions, one synchronous and one asynchronous call (Ping and PingAndWait in the test) hits both