The server is too fast!
In some iterations, during the time we send the MQTT packet, and the time we start listening for the response, the server has already responded.

Aside from the mqtt_packet library, there is a switch statement, which could be done beforehand, and a WaitGroup.Add. The switch statement is not expensive, but moving the Add reduces the number of instructions so the client can start listening for the response before the server has a chance to send it.

Of course, adding a delay in the server would also solve the problem.