The server is too fast!
In some iterations, during the time we send the MQTT packet, and the time we start listening for the response, the server has already responded.

The client spends too much time calling WaitGroup.Add(). This delay means that the goroutine listening for the server response will write to an unbuffered channel before the client is listening to it.

So, for the synchronous calls (connection, subscribe, and unsubscribe), the response channel now has a buffer of 1. So, if the response is written to the channel before we can read it, it won't get dropped into the error queue.